LW-38195: Adds workflow for versioning and publishing
Introduces a new GitHub Actions workflow for building and publishing NuGet packages with manual versioning.
The workflow automates the process of retrieving the version from the project file, building, packing, and publishing the NuGet package to GitHub Packages.
It also includes a tag release job to tag the release with the new version.

Removes the old trunk workflow.